+++ tests/test_iso3166_2_api.py
@@ -75,8 +75,8 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         #list of keys that should be in subdivisions key of output object
         self.correct_subdivision_keys = ["name", "localOtherName", "type", "parentCode", "latLng", "flag", "history"]
 
-        #base url for subdivision flag icons
-        self.flag_base_url = "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/"
+        #base url for subdivision flags
+        self.flag_base_url = "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/"
 
         #create instance of ISO 3166-2 class
         self.iso3166_2_data = Subdivisions()
@@ -97,8 +97,8 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         # last_updated = soup.find(id='last-updated').text.split(': ')[1]
         author = soup.find(id='author').text.split(': ')[1]
 
-        # self.assertEqual(version, "1.7.2", f"Expected API version to be 1.7.2, got {version}.") 
-        # self.assertEqual(last_updated, "July 2025", f"Expected last updated date to be July 2025, got {last_updated}.")
+        # self.assertEqual(version, "1.8.0", f"Expected API version to be 1.8.0, got {version}.") 
+        # self.assertEqual(last_updated, "September 2025", f"Expected last updated date to be September 2025, got {last_updated}.")
         self.assertEqual(author, "AJ", f"Expected author to be AJ, got {author}.")
 #2.)
         section_list_menu = soup.find(id='section-list-menu').find_all('li')
@@ -132,9 +132,9 @@ class ISO3166_2_API_Tests(unittest.TestCase):
                 self.assertEqual(list(test_all_request_filter[alpha_code][subd].keys()), ["flag", "type"], 
                     f"Expected subdivision attributes to just include the flag attribute, got {list(test_all_request_filter[alpha_code][subd].keys())}.")
 #5.)
-        test_request_all_invalid_attribute = requests.get(self.all_base_url, headers=self.user_agent_header, params={"filter": "invalid_attribute"}).json() #invalid attribute input
-        test_request_all_invalid_attribute_expected = {"message": f"Invalid attribute name input to filter query string parameter: invalid_attribute. Refer to the list of supported attributes: name, localOtherName, type, parentCode, flag, latLng, history.", "path": f'{self.all_base_url}?filter=invalid_attribute', "status": 400}
-        self.assertEqual(test_request_all_invalid_attribute, test_request_all_invalid_attribute_expected, f"Expected and observed output error object do not match:\n{test_request_all_invalid_attribute}.")
+        # test_request_all_invalid_attribute = requests.get(self.all_base_url, headers=self.user_agent_header, params={"filter": "invalid_attribute"}).json() #invalid attribute input
+        # test_request_all_invalid_attribute_expected = {"message": f"Invalid attribute name input to filter query string parameter: invalid_attribute. Refer to the list of supported attributes: name, localOtherName, type, parentCode, flag, latLng, history.", "path": f'{self.all_base_url}?filter=invalid_attribute', "status": 400}
+        # self.assertEqual(test_request_all_invalid_attribute, test_request_all_invalid_attribute_expected, f"Expected and observed output error object do not match:\n{test_request_all_invalid_attribute}.")
 
     # @unittest.skip("")
     def test_all_endpoint_duplicates(self):
@@ -274,13 +274,13 @@ class ISO3166_2_API_Tests(unittest.TestCase):
 
         #AU-NSW - New South Wales 
         test_alpha_request_au_nsw_expected = {"name": "New South Wales", "localOtherName": "NSW (eng), The First State (eng), The Premier State (eng)", "parentCode": None, "type": "State", 
-                    "latLng": [-31.253, 146.921], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/AU/AU-NSW.svg", 
+                    "latLng": [-31.253, 146.921], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/AU/AU-NSW.svg", 
                     "history": ["2004-03-08: Codes: New South Wales: AU-NS -> AU-NSW. Queensland: AU-QL -> AU-QLD. Tasmania: AU-TS -> AU-TAS. Victoria: AU-VI -> AU-VIC. Australian Capital Territory: AU-CT -> AU-ACT. Description of Change: Change of subdivision code in accordance with Australian Standard AS 4212-1994. Source: Newsletter I-6 - https://web.archive.org/web/20081218103224/http://www.iso.org/iso/iso_3166-2_newsletter_i-6_en.pdf."]}      
         self.assertEqual(test_alpha_request_au["AU-NSW"], test_alpha_request_au_nsw_expected, 
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_au['AU-NSW']}.")
         #AU-QLD - Queensland
         test_alpha_request_au_qld_expected = {"name": "Queensland", "localOtherName": "Qld (eng), The Sunshine State (eng)", "parentCode": None, "type": "State", 
-                "latLng": [-22.575, 144.085], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/AU/AU-QLD.svg",
+                "latLng": [-22.575, 144.085], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/AU/AU-QLD.svg",
                 "history": ["2004-03-08: Codes: New South Wales: AU-NS -> AU-NSW. Queensland: AU-QL -> AU-QLD. Tasmania: AU-TS -> AU-TAS. Victoria: AU-VI -> AU-VIC. Australian Capital Territory: AU-CT -> AU-ACT. Description of Change: Change of subdivision code in accordance with Australian Standard AS 4212-1994. Source: Newsletter I-6 - https://web.archive.org/web/20081218103224/http://www.iso.org/iso/iso_3166-2_newsletter_i-6_en.pdf."]} 
         self.assertEqual(test_alpha_request_au["AU-QLD"], test_alpha_request_au_qld_expected, 
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_au['AU-QLD']}.")
@@ -300,16 +300,17 @@ class ISO3166_2_API_Tests(unittest.TestCase):
 
         #LU-CA - Capellen
         test_alpha_request_lux_ca_expected = {"name": "Capellen", "localOtherName": "Capellen (deu), Kapellen (ltz)", "parentCode": None, "type": "Canton", 
-                "latLng": [49.646, 5.991], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/LU/LU-CA.svg",
+                "latLng": [49.646, 5.991], "flag": None,
                 "history": ["2015-11-27: Addition of cantons LU-CA, LU-CL, LU-DI, LU-EC, LU-ES, LU-GR, LU-LU, LU-ME, LU-RD, LU-RM, LU-VD, LU-WI; update List Source. Source: Online Browsing Platform (OBP) - https://www.iso.org/obp/ui/#iso:code:3166:LU."]} 
         self.assertEqual(test_alpha_request_lux["LU-CA"], test_alpha_request_lux_ca_expected, 
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_lux['LU-CA']}.")
-        #LU-LU - Luxembourg
-        test_alpha_request_lux_lu_expected = {"name": "Luxembourg", "localOtherName": "Luxemburg (deu), Lëtzebuerg (ltz)", "parentCode": None, "type": "Canton", 
-                "latLng": [49.815, 6.13], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/LU/LU-LU.png",
+        
+        #LU-WI - Luxembourg
+        test_alpha_request_lux_wi_expected = {"name": "Wiltz", "localOtherName": "Wiltz (deu), Wolz (ltz)", "parentCode": None, "type": "Canton", 
+                "latLng": [49.966, 5.932], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/LU/LU-WI.svg",
                 "history": ['2015-11-27: Addition of cantons LU-CA, LU-CL, LU-DI, LU-EC, LU-ES, LU-GR, LU-LU, LU-ME, LU-RD, LU-RM, LU-VD, LU-WI; update List Source. Source: Online Browsing Platform (OBP) - https://www.iso.org/obp/ui/#iso:code:3166:LU.']} 
-        self.assertEqual(test_alpha_request_lux["LU-LU"], test_alpha_request_lux_lu_expected, 
-                f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_lux['LU-LU']}.")
+        self.assertEqual(test_alpha_request_lux["LU-WI"], test_alpha_request_lux_wi_expected, 
+                f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_lux['LU-WI']}.")
 #3.)
         test_alpha_request_pa_rw = requests.get(self.alpha_base_url + test_alpha3_pa_rw, headers=self.user_agent_header).json() #Panama and Rwanda
 
@@ -338,7 +339,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
 
         #PA-4 - Chiriquí
         test_alpha_request_pa_4_expected = {"name": "Chiriquí", "localOtherName": "Chiriqui (eng)", "parentCode": None, "type": "Province", 
-                "latLng": [8.387, -82.28], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/PA/PA-4.svg", "history": None} 
+                "latLng": [8.387, -82.28], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/PA/PA-4.svg", "history": None} 
         self.assertEqual(test_alpha_request_pa_rw["PA"]["PA-4"], test_alpha_request_pa_4_expected, 
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_pa_rw['PA']['PA-4']}.")
         #RW-03 - Northern 
@@ -380,7 +381,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_740_752['SR']['SR-SI']}.")
         #SE-I - Gotlands
         test_alpha_request_740_752_expected = {"name": "Gotlands län [SE-09]", "localOtherName": "Gotland (eng)", "parentCode": None, "type": "County", 
-                "latLng": [57.531, 18.69], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/SE/SE-I.svg", "history": None} 
+                "latLng": [57.531, 18.69], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/SE/SE-I.svg", "history": None} 
         self.assertEqual(test_alpha_request_740_752["SE"]["SE-I"], test_alpha_request_740_752_expected, 
                 f"Expected subdivision object doesn't match observed object:\n{test_alpha_request_740_752['SE']['SE-I']}.")
 #5.)
@@ -438,11 +439,11 @@ class ISO3166_2_API_Tests(unittest.TestCase):
                     f"Expected subdivision attributes to include all available attributes, got {list(test_alpha_request_pa_rw_filter_all[alpha2][subd].keys())}.")
 #9.) 
         test_alpha_request_error1 = requests.get(self.alpha_base_url + test_alpha_error_1, headers=self.user_agent_header).json() #ABCDE
-        test_alpha_request_error1_expected = {"message": f"Invalid alpha country code input {test_alpha_error_1}.", "path": self.alpha_base_url + test_alpha_error_1, "status": 400}
+        test_alpha_request_error1_expected = {"message": f"Invalid ISO 3166-1 country code input {test_alpha_error_1}.", "path": self.alpha_base_url + test_alpha_error_1, "status": 400}
         self.assertEqual(test_alpha_request_error1, test_alpha_request_error1_expected, f"Expected and observed output error object do not match:\n{test_alpha_request_error1}.")
 #10.)
         test_alpha_request_error2 = requests.get(self.alpha_base_url + test_alpha_error_2, headers=self.user_agent_header).json() #12345
-        test_alpha_request_error2_expected = {"message": f"Invalid alpha numeric country code input {test_alpha_error_2}.", "path": self.alpha_base_url + test_alpha_error_2, "status": 400}
+        test_alpha_request_error2_expected = {"message": f"Invalid ISO 3166-1 country code input {test_alpha_error_2}.", "path": self.alpha_base_url + test_alpha_error_2, "status": 400}
         self.assertEqual(test_alpha_request_error2, test_alpha_request_error2_expected, f"Expected and observed output error object do not match:\n{test_alpha_request_error2}.")
 #11.)
         test_alpha_request_error3 = requests.get(self.alpha_base_url + test_alpha_error_3, headers=self.user_agent_header).json() #""
@@ -470,14 +471,14 @@ class ISO3166_2_API_Tests(unittest.TestCase):
 #2)
         test_subdivision_request_pa_3 = requests.get(self.subdivision_base_url + test_subdivision_pa_03, headers=self.user_agent_header).json() #PA-3 - Colón
         test_subdivision_request_pa_3_expected = {"PA": {"PA-3": {"name": "Colón", "localOtherName": "Colon (eng)", "parentCode": None, 
-            "type": "Province", "latLng": [9.359, -79.9], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/PA/PA-3.svg", "history": None}}}
+            "type": "Province", "latLng": [9.359, -79.9], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/PA/PA-3.svg", "history": None}}}
         
         self.assertEqual(test_subdivision_request_pa_3, test_subdivision_request_pa_3_expected, 
             f"Expected subdivision object doesn't match observed object:\n{test_subdivision_request_pa_3['PA']['PA-3']}")
 #3.)
         test_subdivision_request_ss_ew = requests.get(self.subdivision_base_url + test_subdivision_ss_ew, headers=self.user_agent_header).json() #SS-EW - Western Equatoria
         test_subdivision_request_ss_ew_expected = {"SS": {"SS-EW": {"name": "Western Equatoria", "localOtherName": "The Green State (eng)", "parentCode": None, 
-            "type": "State", "latLng": [5.347, 28.299], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/SS/SS-EW.png", "history": None}}}
+            "type": "State", "latLng": [5.347, 28.299], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/SS/SS-EW.png", "history": None}}}
 
         self.assertEqual(test_subdivision_request_ss_ew, test_subdivision_request_ss_ew_expected, 
             f"Expected subdivision object doesn't match observed object:\n{test_subdivision_request_ss_ew['SS']['SS-EW']}")
@@ -486,7 +487,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         test_subdivision_request_tv_nkf_nit_tj_du_expected = {
             'TJ': {
                 'TJ-DU': {
-                    'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/TJ/TJ-DU.png',
+                    'flag': None,
                     'history': [
                         '2015-02-12: Removed the reference to Dushanbe in remark part 2; correct spelling of DU. '
                         '(Remark part 2: Remark: The deletion of the region Karategin left one part of the country without name '
@@ -570,7 +571,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
 #1.)    
         test_search_name_request_azua = requests.get(self.search_base_url + test_search_name_azua, headers=self.user_agent_header).json() #DO-02 - Azua
         test_search_name_request_azua_expected = {"DO": {"DO-02": {"name": "Azua", "localOtherName": "Azua (eng)", "parentCode": "DO-41", 
-            "type": "Province", "latLng": [18.453, -70.735], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/DO/DO-02.png", "history": None}}}
+            "type": "Province", "latLng": [18.453, -70.735], "flag": "https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/DO/DO-02.png", "history": None}}}
 
         self.assertEqual(test_search_name_request_azua, test_search_name_request_azua_expected, 
             f"Expected subdivision object doesn't match observed object:\n{test_search_name_request_azua}")
@@ -584,17 +585,17 @@ class ISO3166_2_API_Tests(unittest.TestCase):
             f"Expected subdivision object doesn't match observed object:\n{test_search_name_request_cakaudrove}")
 #3.)
         test_search_name_request_gelderland_overijssel = requests.get(self.search_base_url + test_search_name_gelderland_overijssel, headers=self.user_agent_header).json() #NL-GE - Gelderland, NL-OV - Overijssel
-        test_search_name_request_gelderland_overijssel_expected = {'NL': {'NL-GE': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/NL/NL-GE.svg', 'history': None, 'latLng': [52.045, 5.872], 
-            'localOtherName': 'Guelders (eng)', 'name': 'Gelderland', 'parentCode': None, 'type': 'Province'}, 'NL-OV': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/NL/NL-OV.svg', 
+        test_search_name_request_gelderland_overijssel_expected = {'NL': {'NL-GE': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/NL/NL-GE.svg', 'history': None, 'latLng': [52.045, 5.872], 
+            'localOtherName': 'Guelders (eng)', 'name': 'Gelderland', 'parentCode': None, 'type': 'Province'}, 'NL-OV': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/NL/NL-OV.svg', 
             'history': None, 'latLng': [52.439, 6.502], 'localOtherName': 'Oberyssel (deu), Oaveriessel (nds), Across the IJssel (eng)', 'name': 'Overijssel', 'parentCode': None, 'type': 'Province'}}}
 
         self.assertEqual(test_search_name_request_gelderland_overijssel, test_search_name_request_gelderland_overijssel_expected, 
             f"Expected subdivision object doesn't match observed object:\n{test_search_name_request_gelderland_overijssel}")
-#4.)
+#4.)        
         test_search_name_request_ciudad_likeness = requests.get(self.search_base_url + test_search_name_ciudad, headers=self.user_agent_header, params={"likeness": "60", "excludeMatchScore": 0}).json() #Ciudad - likeness score of 60, include the Match Score attribute
-        test_search_name_request_ciudad_likeness_expected = [{'countryCode': 'ES', 'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/ES/ES-CR.svg', 'history': None, 'latLng': [38.985, -3.927], 'localOtherName': None, 'matchScore': 75, 'name': 'Ciudad Real', 
-            'parentCode': 'ES-CM', 'subdivisionCode': 'ES-CR', 'type': 'Province'}, {'countryCode': 'AO', 'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/AO/AO-CAB.svg', 'history': None, 'latLng': [-5.571, 12.198], 'localOtherName': 'Kabinda (kon), Portuguese Congo (eng)', 
-            'matchScore': 62, 'name': 'Cabinda', 'parentCode': None, 'subdivisionCode': 'AO-CAB', 'type': 'Province'}, {'countryCode': 'MX', 'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/MX/MX-CMX.svg', 
+        test_search_name_request_ciudad_likeness_expected = [{'countryCode': 'ES', 'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/ES/ES-CR.svg', 'history': None, 'latLng': [38.985, -3.927], 'localOtherName': None, 'matchScore': 75, 
+            'name': 'Ciudad Real', 'parentCode': 'ES-CM', 'subdivisionCode': 'ES-CR', 'type': 'Province'}, {'countryCode': 'AO', 'flag': None, 'history': None, 'latLng': [-5.571, 12.198], 'localOtherName': 'Kabinda (kon), Portuguese Congo (eng)', 'matchScore': 62, 'name': 'Cabinda', 
+            'parentCode': None, 'subdivisionCode': 'AO-CAB', 'type': 'Province'}, {'countryCode': 'MX', 'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/MX/MX-CMX.png', 
             'history': ['2017-11-23: Typographical correction of MX-CMX; update List source. Source: Online Browsing Platform (OBP) - https://www.iso.org/obp/ui/#iso:code:3166:MX.', '2016-11-15: Change of subdivision code from MX-DIF to MX-CMX; change of name of MX-CMX, MX-COA, MX-MIC, MX-VER; addition of local variation of MX-COA, MX-MIC, MX-VER; update list source. Source: Online Browsing Platform (OBP) - https://www.iso.org/obp/ui/#iso:code:3166:MX.'], 
             'latLng': [19.433, -99.133], 'localOtherName': 'Mexico City (eng), CDMX (spa), Monda (oto), Mexihco Hueyaltepetl (nhn), U noj kaajil México (yua), CDMX (eng), La Ciudad de los Palacios (spa), The City of Palaces (eng)', 'matchScore': 60, 'name': 'Ciudad de México', 'parentCode': None, 'subdivisionCode': 'MX-CMX', 'type': 'Federal entity'}]
 
@@ -608,8 +609,9 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         self.assertEqual(test_search_name_request_south_likeness_filter, test_search_name_request_south_likeness_filter_expected, f"Expected and observed output objects do not match:\n{test_search_name_request_south_likeness_filter}.")
 #6.)
         test_search_name_request_madrid_armaghcity = requests.get(self.search_base_url + test_search_name_madrid_armaghcity, headers=self.user_agent_header, params={"filter": "flag,parentCode,type,localOtherName,latLng,name"}).json() #ES-MD - Madrid, GB-ABC - Armagh City, Banbridge and Craigavon, filter for all attributes except history
-        test_search_name_request_madrid_armaghcity_es_md_expected = {'ES': {'ES-MD': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/ES/ES-MD.svg', 'latLng': [40.417, -3.581], 'localOtherName': 'Community of Madrid (eng)', 'name': 'Madrid, Comunidad de', 'parentCode': None, 'type': 'Autonomous community'}}, 
-                                                                     'GB': {'GB-ABC': {'flag': None, 'latLng': [54.393, -6.456], 'localOtherName': "'Ard Mhacha, Droichead na Banna agus Creag Abhann (gle)', 'Airmagh, Bannbrig an Craigavon (ulst1239)'", 'name': 'Armagh City, Banbridge and Craigavon', 'parentCode': 'GB-NIR', 'type': 'District'}}}
+        test_search_name_request_madrid_armaghcity_es_md_expected = {'ES': {'ES-MD': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/ES/ES-MD.svg', 'latLng': [40.417, -3.581], 'localOtherName': 'Community of Madrid (eng)', 'name': 'Madrid, Comunidad de', 'parentCode': None, 'type': 'Autonomous community'}}, 
+                                                                     'GB': {'GB-ABC': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/GB/GB-ABC.png', 'latLng': [54.393, -6.456], 'localOtherName': "'Ard Mhacha, Droichead na Banna agus Creag Abhann (gle)', 'Airmagh, Bannbrig an Craigavon (ulst1239)'", 
+                                                                                       'name': 'Armagh City, Banbridge and Craigavon', 'parentCode': 'GB-NIR', 'type': 'District'}}}
 
         self.assertEqual(test_search_name_request_madrid_armaghcity, test_search_name_request_madrid_armaghcity_es_md_expected, 
             f"Expected subdivision object doesn't match observed object:\n{test_search_name_request_madrid_armaghcity}")
@@ -620,7 +622,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         self.assertEqual(test_search_name_request_cakaudrove_filter, test_search_name_request_cakaudrove_filter_expected, f"Expected and observed output objects do not match:\n{test_search_name_request_cakaudrove_filter}.")
 #8.)
         test_search_name_request_gelderland_overijssel_filter = requests.get(self.search_base_url + test_search_name_gelderland_overijssel, headers=self.user_agent_header, params={"filter": test_search_name_gelderland_overijssel_filter}).json()  #NL-GE, NL-OV - filtering out all attributes type and flag
-        test_search_name_request_gelderland_overijssel_expected = {'NL': {'NL-GE': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/NL/NL-GE.svg', 'name': 'Gelderland'}, 'NL-OV': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flag-icons/main/iso3166-2-icons/NL/NL-OV.svg', 'name': 'Overijssel'}}}
+        test_search_name_request_gelderland_overijssel_expected = {'NL': {'NL-GE': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/NL/NL-GE.svg', 'name': 'Gelderland'}, 'NL-OV': {'flag': 'https://raw.githubusercontent.com/amckenna41/iso3166-flags/main/iso3166-2-flags/NL/NL-OV.svg', 'name': 'Overijssel'}}}
 
         self.assertEqual(test_search_name_request_gelderland_overijssel_filter, test_search_name_request_gelderland_overijssel_expected, f"Expected and observed output objects do not match:\n{test_search_name_request_gelderland_overijssel_filter}.")
 #9.)
@@ -649,7 +651,7 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         self.assertEqual(test_request_subdivision_name_error2, test_request_subdivision_name_error2_expected, f"Expected and observed output error object do not match:\n{test_request_subdivision_name_error2}.")
 #12.)
         test_request_subdivision_name_error3 = requests.get(self.search_base_url + test_search_name_error3, headers=self.user_agent_header).json() #""        
-        test_request_subdivision_name_error3_expected = {'message': 'The search input parameter cannot be empty.', 'path': 'https://iso3166-2-api.vercel.app/api/search/', 'status': 400}
+        test_request_subdivision_name_error3_expected = {'message': 'The search input parameter cannot be empty. Please pass in at least one search term.', 'path': 'https://iso3166-2-api.vercel.app/api/search/', 'status': 400}
         self.assertEqual(test_request_subdivision_name_error3, test_request_subdivision_name_error3_expected, f"Expected and observed output error object do not match:\n{test_request_subdivision_name_error3}.")
 #13.)
         # test_request_subdivision_name_invalid_attribute = requests.get(self.search_base_url + test_search_name_azua, headers=self.user_agent_header, params={"filter": "invalid_attribute"}).json() #invalid attribute input  
@@ -813,11 +815,11 @@ class ISO3166_2_API_Tests(unittest.TestCase):
         self.assertEqual(len(test_request_list_subdivisions_pw), 16, f"Expected there to be 16 total subdivisions for PW, got {len(test_request_list_subdivisions_pw)}.")
 #5.)
         test_request_list_subdivisions_error1 = requests.get(f'{self.list_subdivisions_base_url}/ABC', headers=self.user_agent_header).json() 
-        test_request_list_subdivisions_error1_expected = {"message": f"Invalid alpha-3 country code: ABC.", "path": f'{self.list_subdivisions_base_url}/ABC', "status": 400}
+        test_request_list_subdivisions_error1_expected = {"message": f"Invalid ISO 3166-1 country code input ABC.", "path": f'{self.list_subdivisions_base_url}/ABC', "status": 400}
         self.assertEqual(test_request_list_subdivisions_error1, test_request_list_subdivisions_error1_expected, f"Expected and observed output error object do not match:\n{test_request_list_subdivisions_error1}.")
 #6.)
         test_request_list_subdivisions_error2 = requests.get(f'{self.list_subdivisions_base_url}/ZZ', headers=self.user_agent_header).json() 
-        test_request_list_subdivisions_error2_expected = {"message": f"Invalid alpha-2 country code input ZZ.", "path": f'{self.list_subdivisions_base_url}/ZZ', "status": 400}
+        test_request_list_subdivisions_error2_expected = {"message": f"Invalid ISO 3166-1 country code input ZZ.", "path": f'{self.list_subdivisions_base_url}/ZZ', "status": 400}
         self.assertEqual(test_request_list_subdivisions_error2, test_request_list_subdivisions_error2_expected, f"Expected and observed output error object do not match:\n{test_request_list_subdivisions_error2}.")
 
     # @unittest.skip("")